requirements.txt
requests
tensorflow-gpu
lxml
pyunpack
patool